Cleanup for import PR
